Added portable C bindings to the `fusion_io` Fortran module.
The Fortran `*_f` versions of the functions now directly call the C API
in a portable way.
This has been accomplished by adding an interface block in `fusion_io`
with a corresponding interface for each C API function that has been
marked with the `BIND(C)` identifier and uses C types with value
semantics where appropriate.
The `fortran_interface.cpp` file is no longer needed.
There is some possibility that this will increase performance by fewer
steps of indirection.

It may be possible to add a Fortran interface for
`fio_get_available_fields` as well.